Move 'ChannelModeExtensions' out to its own file
Move `ChannelModeExtensions` out to its own file, located along
with the other 'Extensions classes.
